refactor: use .graphql extension for group mutations
