--- src/apollo/groups/mutations/CancelMemberRequest.mutation.ts
@@ -1,7 +1,0 @@
-import { gql } from "@apollo/client";
-
-gql`
-  mutation CancelMemberRequest($id: Int!) {
-    cancelMemberRequest(id: $id)
-  }
-`;
