+++ src/apollo/groups/mutations/CreateGroup.mutation.graphql
@@ -1,0 +1,11 @@
+mutation CreateGroup($groupData: CreateGroupInput!) {
+  createGroup(groupData: $groupData) {
+    group {
+      ...GroupAvatar
+      description
+      members {
+        ...CurrentMember
+      }
+    }
+  }
+}
