+++ src/apollo/groups/mutations/CreateMemberRequest.mutation.graphql
@@ -1,0 +1,14 @@
+mutation CreateMemberRequest($groupId: Int!) {
+  createMemberRequest(groupId: $groupId) {
+    memberRequest {
+      id
+      group {
+        id
+        name
+      }
+      user {
+        ...UserAvatar
+      }
+    }
+  }
+}
