--- src/apollo/groups/mutations/CreateMemberRequest.mutation.ts
@@ -1,20 +1,0 @@
-import { gql } from "@apollo/client";
-import { UserAvatarFragmentDoc } from "../../gen";
-
-gql`
-  mutation CreateMemberRequest($groupId: Int!) {
-    createMemberRequest(groupId: $groupId) {
-      memberRequest {
-        id
-        group {
-          id
-          name
-        }
-        user {
-          ...UserAvatar
-        }
-      }
-    }
-  }
-  ${UserAvatarFragmentDoc}
-`;
