--- src/apollo/groups/mutations/DeleteGroup.mutation.ts
@@ -1,25 +1,0 @@
-import { ApolloCache, gql } from "@apollo/client";
-import produce from "immer";
-import { TypeNames } from "../../../constants/common.constants";
-import { GroupsDocument, GroupsQuery } from "../../gen";
-
-export const removeGroup = (id: number) => (cache: ApolloCache<any>) => {
-  cache.updateQuery<GroupsQuery>({ query: GroupsDocument }, (groupsData) =>
-    produce(groupsData, (draft) => {
-      if (!draft) {
-        return;
-      }
-      const index = draft.groups.findIndex((p) => p.id === id);
-      draft.groups.splice(index, 1);
-    })
-  );
-  const cacheId = cache.identify({ id, __typename: TypeNames.Group });
-  cache.evict({ id: cacheId });
-  cache.gc();
-};
-
-gql`
-  mutation DeleteGroup($id: Int!) {
-    deleteGroup(id: $id)
-  }
-`;
