+++ src/apollo/groups/mutations/DenyMemberRequest.mutation.graphql
@@ -1,0 +1,3 @@
+mutation DenyMemberRequest($id: Int!) {
+  denyMemberRequest(id: $id)
+}
