--- src/apollo/groups/mutations/DenyMemberRequest.mutation.ts
@@ -1,7 +1,0 @@
-import { gql } from "@apollo/client";
-
-gql`
-  mutation DenyMemberRequest($id: Int!) {
-    denyMemberRequest(id: $id)
-  }
-`;
