+++ src/apollo/groups/mutations/LeaveGroup.mutation.graphql
@@ -1,0 +1,3 @@
+mutation LeaveGroup($id: Int!) {
+  leaveGroup(id: $id)
+}
