--- src/apollo/groups/mutations/LeaveGroup.mutation.ts
@@ -1,7 +1,0 @@
-import { gql } from "@apollo/client";
-
-gql`
-  mutation LeaveGroup($id: Int!) {
-    leaveGroup(id: $id)
-  }
-`;
