+++ src/apollo/groups/mutations/UpdateGroup.mutation.graphql
@@ -1,0 +1,9 @@
+mutation UpdateGroup($groupData: UpdateGroupInput!) {
+  updateGroup(groupData: $groupData) {
+    group {
+      id
+      name
+      description
+    }
+  }
+}
