--- src/apollo/groups/mutations/UpdateGroup.mutation.ts
@@ -1,13 +1,0 @@
-import { gql } from "@apollo/client";
-
-gql`
-  mutation UpdateGroup($groupData: UpdateGroupInput!) {
-    updateGroup(groupData: $groupData) {
-      group {
-        id
-        name
-        description
-      }
-    }
-  }
-`;
